Create input_buffer
Creating the code document for the input buffer. It will contain 2 modules: One pre-register for taking data from the 32-bit BRAM bus. The data must then be formatted as 8 wires with 8 bits to be inserted into the systolic array processing elements. There will be a different document for the controller, which should control both the input and weight buffer. The input buffer and the weight buffer will be the exact same architecture, with the only difference being the data that is passing through.